+++ src/net/_5tingray/stingUtils/color/Color4f.java
@@ -47,12 +47,12 @@ public class Color4f implements ColorI<Color4f> {
 	
 	@Override
 	public Color3i toColor3i() {
-		return new Color3i((int) Math.round(r * 255), (int) Math.round(g * 255), (int) Math.round(b * 255));
+		return new Color3i(Math.round(r * 255), Math.round(g * 255), Math.round(b * 255));
 	}
 
 	@Override
 	public Color4i toColor4i() {
-		return new Color4i((int) Math.round(r * 255), (int) Math.round(g * 255), (int) Math.round(b * 255), (int) Math.round(a * 255));
+		return new Color4i(Math.round(r * 255), Math.round(g * 255), Math.round(b * 255), Math.round(a * 255));
 	}
 
 	@Override
@@ -74,22 +74,22 @@ public class Color4f implements ColorI<Color4f> {
 
 	@Override
 	public int getRedInt() {
-		return (int) Math.round(r * 255);
+		return Math.round(r * 255);
 	}
 
 	@Override
 	public int getGreenInt() {
-		return (int) Math.round(g * 255);
+		return Math.round(g * 255);
 	}
 
 	@Override
 	public int getBlueInt() {
-		return (int) Math.round(b * 255);
+		return Math.round(b * 255);
 	}
 
 	@Override
 	public int getAlphaInt() {
-		return (int) Math.round(a * 255);
+		return Math.round(a * 255);
 	}
 
 	@Override
